refactoring frontend code
