+++ app/organizations/[orgId]/documents/page.tsx
@@ -10,10 +10,12 @@ interface DocumentsPageProps {
 
 export default async function DocumentsPage({ params }: DocumentsPageProps) {
   const { orgId } = await params;
+
+  console.log("DocumentsPage debugging 123456 ", orgId)
   
   return (
     <Suspense fallback={<PageSkeleton />}>
-      <DocumentsContent orgId={orgId} />
+      <DocumentsContent id={orgId} />
     </Suspense>
   );
 }
