--- app/questions/components/questions-section.tsx
@@ -1,2 +1,0 @@
-// Export the refactored version
-export { QuestionsSection } from './questions-section-refactored';
