+++ components/organizations/InviteMemberDialog.tsx
@@ -1,0 +1,149 @@
+'use client';
+
+import React, { useState } from "react";
+import { Button } from "@/components/ui/button";
+import { UserPlus } from "lucide-react";
+import { Input } from "@/components/ui/input";
+import { useToast } from "@/components/ui/use-toast";
+import { 
+  Dialog, 
+  DialogContent, 
+  DialogDescription, 
+  DialogFooter, 
+  DialogHeader, 
+  DialogTitle, 
+  DialogTrigger,
+  DialogClose
+} from "@/components/ui/dialog";
+import { Label } from "@/components/ui/label";
+import { TeamMember } from "./types";
+
+interface InviteMemberDialogProps {
+  orgId: string;
+  onMemberAdded: (member: TeamMember) => void;
+}
+
+export function InviteMemberDialog({ orgId, onMemberAdded }: InviteMemberDialogProps) {
+  const [isInviteDialogOpen, setIsInviteDialogOpen] = useState(false);
+  const [inviteEmail, setInviteEmail] = useState("");
+  const [inviteRole, setInviteRole] = useState<"admin" | "member">("member");
+  const [isInviting, setIsInviting] = useState(false);
+  const { toast } = useToast();
+
+  const handleInviteMember = async (event: React.FormEvent) => {
+    event.preventDefault();
+    
+    if (!inviteEmail.trim()) {
+      toast({
+        title: "Error",
+        description: "Email is required",
+        variant: "destructive",
+      });
+      return;
+    }
+    
+    try {
+      setIsInviting(true);
+      
+      const response = await fetch(`/api/organizations/${orgId}/members/invite`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          email: inviteEmail,
+          role: inviteRole,
+        }),
+      });
+      
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error.message || "Failed to invite member");
+      }
+      
+      // Simulate adding a new member for demo purposes
+      const newMember: TeamMember = {
+        id: `tmp-${Date.now()}`,
+        name: inviteEmail.split('@')[0],
+        email: inviteEmail,
+        role: inviteRole,
+        joinedAt: new Date().toISOString(),
+      };
+      
+      onMemberAdded(newMember);
+      
+      toast({
+        title: "Success",
+        description: `Invitation sent to ${inviteEmail}`,
+      });
+      
+      setInviteEmail("");
+      setInviteRole("member");
+      setIsInviteDialogOpen(false);
+    } catch (error) {
+      console.error("Error inviting member:", error);
+      toast({
+        title: "Error",
+        description: error instanceof Error ? error.message : "Failed to invite member",
+        variant: "destructive",
+      });
+    } finally {
+      setIsInviting(false);
+    }
+  };
+
+  return (
+    <Dialog open={isInviteDialogOpen} onOpenChange={setIsInviteDialogOpen}>
+      <DialogTrigger asChild>
+        <Button>
+          <UserPlus className="mr-2 h-4 w-4" />
+          Invite Member
+        </Button>
+      </DialogTrigger>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>Invite Team Member</DialogTitle>
+          <DialogDescription>
+            Send an invitation to join your organization.
+          </DialogDescription>
+        </DialogHeader>
+        <form onSubmit={handleInviteMember}>
+          <div className="grid gap-4 py-4">
+            <div className="grid gap-2">
+              <Label htmlFor="email">Email address</Label>
+              <Input
+                id="email"
+                type="email"
+                value={inviteEmail}
+                onChange={(e) => setInviteEmail(e.target.value)}
+                placeholder="Enter email address"
+                autoComplete="email"
+                required
+              />
+            </div>
+            <div className="grid gap-2">
+              <Label htmlFor="role">Role</Label>
+              <select 
+                id="role"
+                className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                value={inviteRole}
+                onChange={(e) => setInviteRole(e.target.value as "admin" | "member")}
+              >
+                <option value="member">Member</option>
+                <option value="admin">Admin</option>
+              </select>
+            </div>
+          </div>
+          <DialogFooter>
+            <DialogClose asChild>
+              <Button type="button" variant="outline">Cancel</Button>
+            </DialogClose>
+            <Button type="submit" disabled={isInviting}>
+              {isInviting ? "Sending..." : "Send Invitation"}
+            </Button>
+          </DialogFooter>
+        </form>
+      </DialogContent>
+    </Dialog>
+  );
+}
