+++ components/organizations/index.ts
@@ -5,12 +5,21 @@ export { ProjectsList } from './ProjectsList';
 export { ProjectCard } from './ProjectCard';
 export { EmptyProjectsState } from './EmptyProjectsState';
 
+// Team components
+export { TeamContent } from './TeamContent';
+export { TeamHeader } from './TeamHeader';
+export { TeamMembersTable } from './TeamMembersTable';
+export { MemberTableRow } from './MemberTableRow';
+export { MemberActionsDropdown } from './MemberActionsDropdown';
+export { InviteMemberDialog } from './InviteMemberDialog';
+
 // Other existing components
 export { CreateOrganizationDialog } from './CreateOrganizationDialog';
 export { DocumentsContent } from './DocumentsContent';
 export { LlamaCloudConnection } from './LlamaCloudConnection';
 export { LlamaCloudConnectionDialog } from './LlamaCloudConnectionDialog';
 export { LlamaCloudDocuments } from './LlamaCloudDocuments';
-export { OrganizationSelector } from './OrganizationSelector';
 export { SettingsContent } from './SettingsContent';
-export { TeamContent } from './TeamContent'; 
+
+// Types
+export type { TeamMember } from './types';
