+++ components/organizations/types.ts
@@ -1,0 +1,8 @@
+export interface TeamMember {
+  id: string;
+  name: string;
+  email: string;
+  role: 'owner' | 'admin' | 'member';
+  joinedAt: string;
+  avatarUrl?: string;
+}
